Update schedule to reflect burden of earlier testing
We tried in this cohort of Phase II to move
testing earlier.  The extra overhead and the
instruction in the world of testing has come at a
rate-of-delivery cost.

We have been forced to drop several challenges but
are preferring to focus on *quality* of the
delivery.  As such, as of Thursday morning we want
craiglist jr and anonymous blog to be completed
and to have the former on Heroku.

We will orient Thursday to focus on session and
state and we will also include CSS.  I believe
that the majority of the students are
semi-familiar with CSS and would prefer to have
the time focused on sessions and HTTP / Sinatra
/ Statefulness over CSS.

As such we're going to give them a core
requirement for more back-end work and then we'll
let them pick and choose CSS challenges and, if
they decide *those* are of no use, they will be
allowed to do another set of back-end challenges.